Fix offhand dual wielding mixin
