+++ middle-earth/src/main/java/net/sevenstars/middleearth/mixin/LivingEntityMixin.java
@@ -3,11 +3,16 @@ package net.sevenstars.middleearth.mixin;
 import net.minecraft.enchantment.Enchantment;
 import net.minecraft.enchantment.EnchantmentHelper;
 import net.minecraft.entity.*;
+import net.minecraft.entity.player.PlayerInventory;
 import net.minecraft.item.ItemStack;
 import net.minecraft.registry.RegistryKeys;
 import net.minecraft.registry.entry.RegistryEntry;
+import net.minecraft.util.Hand;
 import net.minecraft.world.World;
 import net.sevenstars.middleearth.enchantments.EnchantmentEffectsME;
+import net.sevenstars.middleearth.item.items.shields.CustomSiegeShieldItem;
+import net.sevenstars.middleearth.item.items.weapons.ReachWeaponItem;
+import net.sevenstars.middleearth.item.items.weapons.ranged.CustomLongbowWeaponItem;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -17,10 +22,13 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(LivingEntity.class)
 public abstract class LivingEntityMixin extends Entity {
+    private static boolean stackTraceLock = false;
 
     @Shadow @Final protected EntityEquipment equipment;
 
     @Shadow public abstract ItemStack getEquippedStack(EquipmentSlot slot);
+
+    @Shadow public abstract ItemStack getStackInHand(Hand hand);
 
     public LivingEntityMixin(EntityType<?> type, World world) {
         super(type, world);
@@ -42,5 +50,37 @@ public abstract class LivingEntityMixin extends Entity {
                 }
             }
         }
+    }
+
+    @Inject(method = "getEquippedStack", at = @At("HEAD"), cancellable = true)
+    public void getEquippedStack(EquipmentSlot slot, CallbackInfoReturnable<ItemStack> cir) {
+        if(stackTraceLock) return;
+        stackTraceLock = true;
+        boolean twoHanded = false;
+        ItemStack stackMainHand = this.getStackInHand(Hand.MAIN_HAND);
+        ItemStack stackOffHand = this.getStackInHand(Hand.OFF_HAND);
+
+        if(stackMainHand != null){
+            if ((stackMainHand.getItem() instanceof ReachWeaponItem && (((ReachWeaponItem) stackMainHand.getItem()).type.twoHanded))
+                    || (stackMainHand.getItem() instanceof CustomSiegeShieldItem)
+                    || (stackMainHand.getItem() instanceof CustomLongbowWeaponItem)) {
+                twoHanded = true;
+            }
+        }
+        if(stackOffHand != null){
+            if ((stackOffHand.getItem() instanceof ReachWeaponItem && (((ReachWeaponItem) stackOffHand.getItem()).type.twoHanded))
+                    || (stackOffHand.getItem() instanceof CustomSiegeShieldItem)
+                    || (stackOffHand.getItem() instanceof CustomLongbowWeaponItem)) {
+                twoHanded = true;
+            }
+        }
+
+        if (slot == EquipmentSlot.OFFHAND) {
+            if (twoHanded) {
+                cir.setReturnValue(ItemStack.EMPTY);
+                cir.cancel();
+            }
+        }
+        stackTraceLock = false;
     }
 }
